moves keycodes to constants folder
