+++ components/src/Constants/index.js
@@ -1,0 +1,1 @@
+export { default as keyCodes } from "../Constants/keyCodes";
