+++ components/src/NavBar/withMenuState.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { keyCodes } from "../Utils";
+import { keyCodes } from "../Constants";
 
 class MenuState extends React.Component {
   constructor() {
