+++ components/src/Tooltip/Tooltip.js
@@ -4,7 +4,8 @@ import PropTypes from "prop-types";
 import { Manager, Reference, Popper } from "react-popper";
 import { Box } from "../Box";
 import theme from "../theme";
-import { withGeneratedId, DetectOutsideClick, PopperArrow, keyCodes } from "../Utils";
+import { withGeneratedId, DetectOutsideClick, PopperArrow } from "../Utils";
+import { keyCodes } from "../Constants";
 
 const tooltipStyles = {
   backgroundColor: theme.colors.white,
