+++ components/src/Utils/index.js
@@ -5,4 +5,3 @@ export { default as withGeneratedId } from "./withGeneratedId";
 export { default as withWindowDimensions } from "./withWindowDimensions";
 export { default as DetectOutsideClick } from "./DetectOutsideClick";
 export { default as PopperArrow } from "./PopperArrow";
-export { default as keyCodes } from "./keyCodes";
